backend/requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
python-decouple==3.7
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
uvicorn==0.27.0.post1
psycopg2-binary==2.9.9
fastapi-mail==1.4.2
pytest==8.3.5
python-dotenv==1.1.0
asyncpg==0.30.0
twilio==9.7.0
minio==7.2.16
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.20
prometheus-fastapi-instrumentator==7.1.0
redis==6.4.0
fastapi-cache2==0.2.2
alembic==1.16.5
greenlet==3.3.2
